\name{RcmdrPlugin.steepness-internal}
\title{RcmdrPlugin.steepness: A GUI for steepness-package (internal functions)}
\alias{Rcmdr.help.RcmdrPlugin.steepness}
\alias{Rcmdr.help.steepness}
\alias{Rcmdr.steepplot}
\alias{Rcmdr.steeptestDij}
\alias{Rcmdr.steeptestPij}

\encoding{UTF-8}
\description{Internal functions (should not be called by the user) for
  RcmdrPlugin.steepness package. 
  For user information ask for help about \code{RcmdrPlugin.steepness}.}
\keyword{internal}
