\name{RcmdrPlugin.steepness-package}
\title{RcmdrPlugin.steepness: A GUI for steepness R package}
\alias{RcmdrPlugin.steepness}
\docType{package}
\encoding{UTF-8}
\description{
  This package provides a GUI (Graphical User Interface) for the
  steepness package. It is a Plug-in for Rcmdr (R Commander).

  Some functions of \code{steepness} package are available through the
  GUI.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.steepness\cr
Type: \tab Package\cr
Version: \tab 0.3-2\cr
Date: \tab 2014-29-09\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
David Leiva <dleivaur@ub.edu> & Han de Vries <J.deVries1@uu.nl>

Maintainer: David Leiva <dleivaur@ub.edu>
}

\examples{
  ??RcdmrPlugin.steepness
}

\keyword{package}
\seealso{
For more information see \code{Rcmdr} and \code{steepness}.
}

